Endpoints de plantilla ready.

+++ BussinessLayer/Interfaces/ModuloCampaña/Repository/ICmpPlantillaRepository.cs
@@ -4,8 +4,8 @@ namespace BussinessLayer.Interfaces.ModuloCampaña.Repository
 {
     public interface ICmpPlantillaRepository
     {
-        Task<CmpPlantillas> GetByIdAsync(int id, int empresaId);
-        Task<List<CmpPlantillas>> GetAllAsync(int empresaId);
+        Task<CmpPlantillas> GetByIdAsync(int id, long empresaId);
+        Task<List<CmpPlantillas>> GetAllAsync(long empresaId);
         Task AddAsync(CmpPlantillas plantilla);
         Task DeleteAsync(int id);
         Task UpdateAsync(CmpPlantillas plantilla);
